@@ -1,7 +1,11 @@
+// Package fossilfs provides an experimental storage backend for gopass
+// based on the Fossil SCM system.
+//
+// It is not recommended for production use and is only intended for testing
+// purposes.
 package fossilfs
 
 import (
-	"context"
 	"os"
 	"path/filepath"
 	"testing"
@@ -21,7 +25,7 @@ func createMarker(t *testing.T, path string) {
 
 func TestLoader_New(t *testing.T) {
 	l := loader{}
-	ctx := context.Background()
+	ctx := t.Context()
 	path := t.TempDir()
 	createMarker(t, path)
 
@@ -32,7 +36,7 @@ func TestLoader_New(t *testing.T) {
 
 func TestLoader_Open(t *testing.T) {
 	l := loader{}
-	ctx := context.Background()
+	ctx := t.Context()
 	path := t.TempDir()
 	createMarker(t, path)
 
@@ -45,7 +49,7 @@ func TestLoader_Clone(t *testing.T) {
 	t.Skip("needs fossil binary and valid remote")
 
 	l := loader{}
-	ctx := context.Background()
+	ctx := t.Context()
 	repo := "https://example.com/repo.git"
 	path := t.TempDir()
 	createMarker(t, path)
@@ -59,7 +63,7 @@ func TestLoader_Init(t *testing.T) {
 	t.Skip("needs fossil binary")
 
 	l := loader{}
-	ctx := context.Background()
+	ctx := t.Context()
 	path := t.TempDir()
 	createMarker(t, path)
 
@@ -70,7 +74,7 @@ func TestLoader_Init(t *testing.T) {
 
 func TestLoader_Handles(t *testing.T) {
 	l := loader{}
-	ctx := context.Background()
+	ctx := t.Context()
 	td := t.TempDir()
 
 	err := l.Handles(ctx, td)
