@@ -1,7 +1,6 @@
 package config
 
 import (
-	"context"
 	"testing"
 
 	"github.com/gopasspw/gopass/tests/gptest"
@@ -39,7 +38,7 @@ func TestConfig(t *testing.T) {
 	require.NoError(t, cfg.Set("", "pwgen.xkcd-lang", "de"))
 	assert.Equal(t, "de", cfg.Get("pwgen.xkcd-lang"))
 
-	ctx := cfg.WithConfig(context.Background())
+	ctx := cfg.WithConfig(t.Context())
 	assert.True(t, Bool(ctx, "core.bool"))
 	assert.Equal(t, "foo", String(ctx, "core.string"))
 	assert.Equal(t, 42, Int(ctx, "core.int"))
@@ -117,7 +116,7 @@ func TestOptsMigration(t *testing.T) {
 		require.NoError(t, cfg.Set("", "core.showsafecontent", "true"))
 		assert.True(t, cfg.IsSet("core.showsafecontent"))
 		assert.Equal(t, "true", cfg.root.GetGlobal("core.showsafecontent"))
-		assert.Equal(t, "", cfg.root.GetLocal("core.showsafecontent"))
+		assert.Empty(t, cfg.root.GetLocal("core.showsafecontent"))
 		assert.False(t, cfg.IsSet("core.safecontent"))
 		assert.False(t, cfg.IsSet("show.safecontent"))
 
@@ -128,9 +127,9 @@ func TestOptsMigration(t *testing.T) {
 		cfg2 := New()
 		assert.False(t, cfg2.IsSet("core.showsafecontent"))
 		assert.False(t, cfg2.IsSet("core.safecontent"))
-		assert.Equal(t, "", cfg2.root.GetGlobal("core.showsafecontent"))
+		assert.Empty(t, cfg2.root.GetGlobal("core.showsafecontent"))
 		assert.Equal(t, "true", cfg2.root.GetGlobal("show.safecontent"))
-		assert.Equal(t, "", cfg2.root.GetLocal("show.safecontent"))
+		assert.Empty(t, cfg2.root.GetLocal("show.safecontent"))
 	})
 
 	t.Run("migrated config matches test config", func(t *testing.T) {
@@ -169,7 +168,7 @@ func TestOptsMigration(t *testing.T) {
 		// this will write to the local config because of the <root> arg
 		require.NoError(t, cfg.Set("<root>", "core.showsafecontent", "true"))
 		assert.Equal(t, "true", cfg.root.GetLocal("core.showsafecontent"))
-		assert.Equal(t, "", cfg.root.GetGlobal("core.showsafecontent"))
+		assert.Empty(t, cfg.root.GetGlobal("core.showsafecontent"))
 		assert.False(t, cfg.IsSet("show.safecontent"))
 
 		t.Setenv("GOPASS_CONFIG_NO_MIGRATE", "")
@@ -177,7 +176,7 @@ func TestOptsMigration(t *testing.T) {
 		cfg = New()
 		assert.False(t, cfg.IsSet("core.showsafecontent"))
 		assert.Equal(t, "true", cfg.root.GetLocal("show.safecontent"))
-		assert.Equal(t, "", cfg.root.GetGlobal("show.safecontent"))
+		assert.Empty(t, cfg.root.GetGlobal("show.safecontent"))
 	})
 
 	t.Run("env variable are not migrated", func(t *testing.T) {
@@ -214,15 +213,15 @@ func TestOptsMigration(t *testing.T) {
 		// this will write to the local mount config
 		require.NoError(t, cfg.Set("submount", "core.showsafecontent", "true"))
 		assert.Equal(t, "true", cfg.GetM("submount", "core.showsafecontent"))
-		assert.Equal(t, "", cfg.Get("core.showsafecontent"))
-		assert.Equal(t, "", cfg.GetM("submount", "show.safecontent"))
+		assert.Empty(t, cfg.Get("core.showsafecontent"))
+		assert.Empty(t, cfg.GetM("submount", "show.safecontent"))
 
 		t.Setenv("GOPASS_CONFIG_NO_MIGRATE", "")
 		// we test the migration path
 		cfg = New()
 		assert.False(t, cfg.IsSet("core.showsafecontent"))
 		assert.False(t, cfg.IsSet("show.safecontent"))
-		assert.Equal(t, "", cfg.GetM("submount", "core.showsafecontent"))
+		assert.Empty(t, cfg.GetM("submount", "core.showsafecontent"))
 		assert.Equal(t, "true", cfg.GetM("submount", "show.safecontent"))
 	})
 }
