@@ -9,22 +9,23 @@ import (
 )
 
 func TestGPG(t *testing.T) {
-	t.Parallel()
-
 	if testing.Short() {
 		t.Skip("skipping test in short mode.")
 	}
+
+	td := t.TempDir()
+	t.Setenv("GNUPGHOME", td)
 
 	ctx := config.NewContextInMemory()
 
 	var err error
 	var g *GPG
 
-	assert.Equal(t, "", g.Binary())
+	assert.Empty(t, g.Binary())
 
 	g, err = New(ctx, Config{})
 	require.NoError(t, err)
-	assert.NotEqual(t, "", g.Binary())
+	assert.NotEmpty(t, g.Binary())
 
 	_, err = g.ListRecipients(ctx)
 	require.NoError(t, err)
