@@ -13,5 +13,5 @@ func TestTTY(t *testing.T) {
 	t.Parallel()
 
 	fd0 = "/tmp/foobar"
-	assert.Equal(t, "", TTY())
+	assert.Empty(t, TTY())
 }
