@@ -10,10 +10,10 @@ import (
 )
 
 // Fsck checks all stores/entries matching the given prefix.
-func (s *Store) Fsck(ctx context.Context, store, path string) error {
+func (r *Store) Fsck(ctx context.Context, store, path string) error {
 	var result []error
 
-	for alias, sub := range s.mounts {
+	for alias, sub := range r.mounts {
 		if sub == nil {
 			continue
 		}
@@ -41,7 +41,7 @@ func (s *Store) Fsck(ctx context.Context, store, path string) error {
 
 	// check root store
 	debug.Log("Checking root store")
-	if err := s.store.Fsck(ctx, path); err != nil {
+	if err := r.store.Fsck(ctx, path); err != nil {
 		out.Errorf(ctx, "fsck failed on root store: %s", err)
 		result = append(result, err)
 	}
