@@ -213,17 +213,11 @@ func (r *Store) checkMounts() error {
 // CleanMountAlias removes all leading and trailing slashes from a mount alias.
 // Note: Slashes inside the alias are valid and will be kept.
 func CleanMountAlias(alias string) string {
-	for {
-		if !strings.HasPrefix(alias, "/") && !strings.HasPrefix(alias, "\\") {
-			break
-		}
+	for strings.HasPrefix(alias, "/") || strings.HasPrefix(alias, "\\") {
 		alias = strings.TrimPrefix(strings.TrimSuffix(alias, "/"), "/")
 		alias = strings.TrimPrefix(strings.TrimSuffix(alias, "\\"), "\\")
 	}
-	for {
-		if !strings.HasSuffix(alias, "/") && !strings.HasSuffix(alias, "\\") {
-			break
-		}
+	for strings.HasSuffix(alias, "/") || strings.HasSuffix(alias, "\\") {
 		alias = strings.TrimSuffix(strings.TrimPrefix(alias, "/"), "/")
 		alias = strings.TrimSuffix(strings.TrimPrefix(alias, "\\"), "\\")
 	}
